+++ packages/api/nestjs-auth/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+
+## 0.1.0
+<sub>2026-08-06</sub>
+
+- [#1558](https://github.com/wisemen-digital/wisemen-core/pull/1558)  *(minor)* Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - Added basic auth for nestjs.
+- [#1558](https://github.com/wisemen-digital/wisemen-core/pull/1558)  *(minor)* Thanks [@Kobe-Kwanten](https://github.com/Kobe-Kwanten)! - add basic auth
+
 ## 0.0.1
 <sub>2026-07-31</sub>